i18n for aria label on clear button. remove lodash.memoize and implement useMemo hook instead - add prop to typeahead to toggle filter methods to enable hook usage for filtering.
