--- assets/src/dashboard/components/typeahead-options/utils/getTypeaheadOptions.js
@@ -1,37 +1,0 @@
-/*
- * Copyright 2020 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * External dependencies
- */
-// eslint-disable-next-line import/no-extraneous-dependencies
-import memoize from 'lodash/memoize'; // only lodash/memoize is added to package dependencies to avoid code bloat
-
-const getTypeaheadOptions = memoize(
-  (items, inputValue) => {
-    return items.filter((item) => {
-      const lowerInputValue = inputValue.toLowerCase();
-
-      return (
-        item.label.toLowerCase().includes(lowerInputValue) ||
-        item.value.toLowerCase().includes(lowerInputValue)
-      );
-    });
-  },
-  (...args) => JSON.stringify(args)
-);
-
-export default getTypeaheadOptions;
