+++ package.json
@@ -29,7 +29,6 @@
     "draft-js-import-html": "^1.4.1",
     "draftjs-filters": "^2.3.0",
     "history": "^4.10.1",
-    "lodash.memoize": "^4.1.2",
     "mousetrap": "^1.6.5",
     "polished": "^3.5.1",
     "prop-types": "^15.7.2",
